refactor : TT-27 refactor logout component
